Moved files out of redundant "ros" subfolder
Before, we had library compnents in other subfolders, but they got moved out.
